Move code into its own files.
